Create README.md
Asset management (AM) is a process by which public entities make and apply decisions about acquiring, developing, operating, maintaining, refurbishing, or disposing of assets cost effectively. The ultimate objective is to provide public services adequately and sustainably. The asset management framework is a system of rules, procedures, and entities that informs and performs the tasks of AM.

Good AM is important for county governments for several interrelated reasons, including service delivery, urban  development, wealth and income, budgeting, transparency and democracy, and management of liabilities.